authorization, gates and etc

+++ app/Http/Controllers/BookController.php
@@ -6,28 +6,34 @@ use App\Http\Requests\StoreBookRequest;
 use App\Http\Requests\UpdateBookRequest;
 use App\Http\Resources\BookResource;
 use App\Models\Book;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BookController extends Controller
 {
     public function index()
     {
-        return BookResource::collection(Book::all());
+        $books = Auth::user()->books()->get();
+        return BookResource::collection($books);
     }
 
     public function store(StoreBookRequest $request)
     {
-        $book = Book::create($request->validated());
+
+        $book = Auth::user()->books()->create($request->validated());
 
         return new BookResource($book);
     }
 
     public function show(Book $book)
     {
+        Gate::authorize('viewOrModify', $book);
         return new BookResource($book);
     }
 
     public function update(UpdateBookRequest $request, Book $book)
     {
+        Gate::authorize('viewOrModify', $book);
         $book->update($request->validated());
 
         return new BookResource($book);
@@ -35,6 +41,7 @@ class BookController extends Controller
 
     public function destroy(Book $book)
     {
+        Gate::authorize('viewOrModify', $book);
         $book->delete();
 
         return response()->noContent();
